WIP User Story 4
In the feature test for opponent choosing a shape we find an html
element containing the message. Then we generate all the possible
messages and check if one of them matches the actual one. It allows us
not to mock any implementation details (choosing a random element) and make our feature test
completely decoupled from implementation.